Move latex_requirement.rb to specs/ folder since it is only used in specs
