--- lib/latexmath/requirement.rb
@@ -1,11 +1,0 @@
-module Latexmath
-  class Requirement
-    def satisfied
-      raise NotImplementedError('abstract method')
-    end
-
-    def cmd
-      raise NotImplementedError('abstract method')
-    end
-  end
-end
